Add purge_on_timeout keyword arg for automatically deleting expired keys (#7)
* Add purge_on_timeout keyword arg for automatically deleting expired keys

Implements #1. You can now pass `purge_on_timeout=true` to
`ExpiringCaches.Cache` constructor, which will result in an asynchronous
task being spawned whenever a key is added to the cache. The task will
wait for the cache's timeout length, and afterwards remove the key-val
from the cache, UNLESS it has already been replaced, in which case
another timer task was spawned to purge the replaced key.

* fix

* update ci

* fix